+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - New function `mention_list_entities` to parse user and list mentions from given text
   - New function `mention_entities` to parse just user mentions from given text
   - New function `hashtag_entities` to parse just hashtags from given text
+  - New function `symbol_entities` to parse just cashtags from given text
 
 ### Changed
 - `Place::contained_within` is now a `Option<Vec<Place>>`, because I had the wrong type for that before